fix(ws_gw_fct_pg2epa_dscenario): update node_id type casting in demand insertion queries

+++ dbmodel/ws/fct/ws_gw_fct_pg2epa_dscenario.sql
@@ -91,10 +91,10 @@ BEGIN
 
 			-- demands for connec related to arcs
 		INSERT INTO temp_t_demand (dscenario_id, feature_id, demand, pattern_id, demand_type, source)
-			SELECT dscenario_id, node_1::integer AS node_id, d.demand/2 as demand, d.pattern_id, demand_type, source FROM temp_t_arc JOIN ve_inp_connec ON temp_t_arc.arc_id = ve_inp_connec.arc_id::text
+			SELECT dscenario_id, node_1 AS node_id, d.demand/2 as demand, d.pattern_id, demand_type, source FROM temp_t_arc JOIN ve_inp_connec ON temp_t_arc.arc_id = ve_inp_connec.arc_id::text
 			JOIN inp_dscenario_demand d ON feature_id = connec_id WHERE dscenario_id IN (SELECT unnest(v_userscenario)) AND pjoint_type in ('ARC', 'CONNEC')
 			UNION ALL
-			SELECT dscenario_id, node_2::integer AS node_id, d.demand/2 as demand, d.pattern_id, demand_type, source  FROM temp_t_arc JOIN ve_inp_connec ON temp_t_arc.arc_id = ve_inp_connec.arc_id::text
+			SELECT dscenario_id, node_2 AS node_id, d.demand/2 as demand, d.pattern_id, demand_type, source  FROM temp_t_arc JOIN ve_inp_connec ON temp_t_arc.arc_id = ve_inp_connec.arc_id::text
 			JOIN inp_dscenario_demand d ON feature_id = connec_id WHERE dscenario_id IN (SELECT unnest(v_userscenario)) AND pjoint_type in ('ARC', 'CONNEC');
